CASMSEC-385: kyverno resources need to be adjusted
Issue: kyverno and kyverno-pre containers memory resources needed to be increased otherwise we hit OOMKilled and the pods failed to start. When that happens, other critical pods fail to start.
Solution: Taking into consideration of higher configuration systems with future Kyverno rule increment and number of pods associated with that, we have increased CPU requests to 5 times and memory limit to 4 times to the current values.